+++ src/Microsoft.Health.Dicom.Core/Modules/ServiceModule.cs
@@ -6,11 +6,13 @@
 using EnsureThat;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Health.Dicom.Core.Configs;
+using Microsoft.Health.Dicom.Core.Extensions;
 using Microsoft.Health.Dicom.Core.Features.ChangeFeed;
 using Microsoft.Health.Dicom.Core.Features.Common;
 using Microsoft.Health.Dicom.Core.Features.Delete;
 using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
 using Microsoft.Health.Dicom.Core.Features.HealthCheck;
+using Microsoft.Health.Dicom.Core.Features.Indexing;
 using Microsoft.Health.Dicom.Core.Features.Query;
 using Microsoft.Health.Dicom.Core.Features.Retrieve;
 using Microsoft.Health.Dicom.Core.Features.Store;
@@ -168,6 +170,8 @@ namespace Microsoft.Health.Dicom.Core.Modules
                      .Scoped()
                      .AsSelf()
                      .AsImplementedInterfaces();
+
+                services.AddScopedDefault<InstanceReindexer>();
             }
         }
     }
