+++ src/Microsoft.Health.Dicom.Functions/Registration/DicomFunctionsServiceCollectionExtensions.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Microsoft.Health.Dicom.Core.Registration;
-using Microsoft.Health.Dicom.Functions.Indexing.Configuration;
+using Microsoft.Health.Dicom.Functions.Configs;
 
 namespace Microsoft.Health.Dicom.Functions.Registration
 {
